backend/requirements.txt
flask
rembg==2.0.55
pillow